#30: Fix travis
Squashed commit of the following:

commit 43c66f3d684a3e5fa4d0ef9f2238c68d60463b74
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 22 09:28:42 2019 +0000

    #30: Fix (????!!!) travis

commit 124483b129fab65ba795442f4547a8e6f03be098
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 22 09:18:06 2019 +0000

    #30: Fix (??!!) travis

commit a686cbf7e829da915998c524fd5d089bd3dacf28
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 22 09:14:33 2019 +0000

    #30: Specify c++11 explicitly

commit acdfe34be8d5bdf3d2ce98a7164fc4a607af65df
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 22 09:12:52 2019 +0000

    #30: Cache SDL build

commit b193e1ee11915d7ac8fb07357fe41087b6171d9a
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 22 08:49:27 2019 +0000

    #30: Add this branch to travis

commit bbf86b655f9f63c8c7703aaffdeb33489e84670b
Author: Dom Williams <me@domwillia.ms>
Date:   Fri Nov 22 08:46:30 2019 +0000

    #30: Add bullet to travis

+++ .travis-install-sdl2.sh
@@ -3,6 +3,13 @@
 
 set -xueo pipefail
 
-wget https://www.libsdl.org/release/SDL2-2.0.9.tar.gz -O sdl2.tar.gz
-tar xzf sdl2.tar.gz
-pushd SDL2-* && ./configure && make -j`nproc` && sudo make install && popd
+# SDL_VERSION=2.0.9
+# SDL_DIR="SDL2-$SDL_VERSION"
+
+if [ ! -f "$SDL_DIR/configure" ]
+then
+	wget https://www.libsdl.org/release/SDL2-2.0.9.tar.gz -O sdl2.tar.gz
+	tar xzf sdl2.tar.gz
+fi
+
+pushd $SDL_DIR && ./configure && make -j`nproc` && sudo make install && popd
